# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TradingView Screener REST API",
    "description": "TradingView-style scanner API with TypeScript-oriented structured responses",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/api/v1/scan/raw": {
      "post": {
        "tags": [
          "scan"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "properties": {
                  "payload": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "options": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "requestOptions": {
                    "type": "object",
                    "additionalProperties": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "required": [
                    "totalCount",
                    "data"
                  ],
                  "properties": {
                    "totalCount": {
                      "type": "number"
                    },
                    "data": {
                      "type": [
                        "null",
                        "array"
                      ],
                      "items": {
                        "type": "object",
                        "additionalProperties": false,
                        "required": [
                          "s",
                          "d"
                        ],
                        "properties": {
                          "s": {
                            "type": "string"
                          },
                          "d": {
                            "type": "array",
                            "items": {}
                          }
                        }
                      }
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "required": [
                    "totalCount",
                    "data"
                  ],
                  "properties": {
                    "totalCount": {
                      "type": "number"
                    },
                    "data": {
                      "type": [
                        "null",
                        "array"
                      ],
                      "items": {
                        "type": "object",
                        "additionalProperties": false,
                        "required": [
                          "s",
                          "d"
                        ],
                        "properties": {
                          "s": {
                            "type": "string"
                          },
                          "d": {
                            "type": "array",
                            "items": {}
                          }
                        }
                      }
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "required": [
                    "totalCount",
                    "data"
                  ],
                  "properties": {
                    "totalCount": {
                      "type": "number"
                    },
                    "data": {
                      "type": [
                        "null",
                        "array"
                      ],
                      "items": {
                        "type": "object",
                        "additionalProperties": false,
                        "required": [
                          "s",
                          "d"
                        ],
                        "properties": {
                          "s": {
                            "type": "string"
                          },
                          "d": {
                            "type": "array",
                            "items": {}
                          }
                        }
                      }
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scan": {
      "post": {
        "tags": [
          "scan"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "properties": {
                  "payload": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "options": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "requestOptions": {
                    "type": "object",
                    "additionalProperties": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "totalCount",
                    "columns",
                    "rawRows",
                    "mappedRows"
                  ],
                  "properties": {
                    "totalCount": {
                      "type": "number"
                    },
                    "columns": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "rawRows": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": false,
                        "required": [
                          "s",
                          "d"
                        ],
                        "properties": {
                          "s": {
                            "type": "string"
                          },
                          "d": {
                            "type": "array",
                            "items": {}
                          }
                        }
                      }
                    },
                    "mappedRows": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "required": [
                    "totalCount",
                    "data"
                  ],
                  "properties": {
                    "totalCount": {
                      "type": "number"
                    },
                    "data": {
                      "type": [
                        "null",
                        "array"
                      ],
                      "items": {
                        "type": "object",
                        "additionalProperties": false,
                        "required": [
                          "s",
                          "d"
                        ],
                        "properties": {
                          "s": {
                            "type": "string"
                          },
                          "d": {
                            "type": "array",
                            "items": {}
                          }
                        }
                      }
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "required": [
                    "totalCount",
                    "data"
                  ],
                  "properties": {
                    "totalCount": {
                      "type": "number"
                    },
                    "data": {
                      "type": [
                        "null",
                        "array"
                      ],
                      "items": {
                        "type": "object",
                        "additionalProperties": false,
                        "required": [
                          "s",
                          "d"
                        ],
                        "properties": {
                          "s": {
                            "type": "string"
                          },
                          "d": {
                            "type": "array",
                            "items": {}
                          }
                        }
                      }
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/symbol-search/v3": {
      "get": {
        "tags": [
          "search"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "query",
            "name": "text",
            "required": true
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "hl",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "country",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "search_type",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "lang",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "exchange",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "domain",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "sort_by_country",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "required": [
                    "symbols"
                  ],
                  "properties": {
                    "symbols_remaining": {
                      "type": "number"
                    },
                    "symbols": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}